requirements.txt

# LLM & AI
ollama>=0.3.3
httpx>=0.27.0
openai-whisper>=20231117
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
torch>=2.5.0
torchvision>=0.20.0
# Grok (xAI) — no official SDK, use httpx for API calls

# OS Control
psutil>=6.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pywinauto>=0.6.8
pynput>=1.7.7
pyperclip>=1.9.0
GPUtil>=1.4.0
pycaw>=20240210
screen-brightness-control>=0.23.0

# File System
watchdog>=4.0.1
PyPDF2>=3.0.1
python-docx>=1.1.2
send2trash>=1.8.3
fpdf2>=2.7.0

# Document & Media Generation
python-pptx>=0.6.23
matplotlib>=3.9.0
qrcode>=7.4.2
gTTS>=2.5.0
markdown>=3.6
feedparser>=6.0.11
icalendar>=6.0.0
pytesseract>=0.3.10

# Web Scraping (Chrome)
selenium>=4.20.0
webdriver-manager>=4.0.1
lxml>=5.2.0

# Server & Async
fastapi>=0.111.1
uvicorn>=0.30.1

# UI & Audio
pyttsx3>=2.90
rich>=13.7.1
plyer>=2.1.0
sounddevice>=0.4.7

# Data
networkx>=3.3
pandas>=2.2.0
openpyxl>=3.1.2
pydantic>=2.8.2
pyyaml>=6.0.1
python-dotenv>=1.0.1
numpy>=1.26.4

# Scheduling
APScheduler>=3.10.4
